indent the "bodies" where appropriate
Even though here we use 2 spaces, yasnippet will automatically indent
each line according to the mode, which takes into account any changes to
the indentation made by the user.

This wasn't happening previously presumably because the line was
considered "empty" and hence there was nothing to indent.

Now that the line is not empty, it will be properly indented.